+++ Sources/SwiftLinkPreview.swift
@@ -10,11 +10,12 @@ import Foundation
 public class SwiftLinkPreview {
     
     // MARK: - Vars
-    static let minimumRelevant: Int = 120
-    private var text: String!
+    static let titleMinimumRelevant: Int = 15
+    static let decriptionMinimumRelevant: Int = 100
     private var url: NSURL!
     private var task: NSURLSessionDataTask?
     private let session = NSURLSession.sharedSession()
+    internal var text: String!
     internal var result: [String: AnyObject] = [:]
     
     // MARK: - Constructor
@@ -39,7 +40,7 @@ public class SwiftLinkPreview {
                 
                 self.result["finalUrl"] = unshortened
                 
-                self.extractCanonicalURL()
+                self.result["canonicalUrl"] = self.extractCanonicalURL(unshortened)
                 
                 self.extractInfo({
                     
@@ -97,23 +98,15 @@ public class SwiftLinkPreview {
 extension SwiftLinkPreview {
     
     // Extract first URL from text
-    private func extractURL() -> NSURL? {
+    internal func extractURL() -> NSURL? {
         
         let explosion = self.text.characters.split{$0 == " "}.map(String.init)
+        let pieces = explosion.filter({ $0.trim.isValidURL() })
+        let piece = pieces[0]
         
-        for var piece in explosion {
+        if let url = NSURL(string: piece) {
             
-            piece = piece.trim
-            
-            if let url = NSURL(string: piece.trim) {
-                
-                if url.absoluteString.isValidURL() {
-                    
-                    return url
-                    
-                }
-                
-            }
+            return url
             
         }
         
@@ -130,7 +123,7 @@ extension SwiftLinkPreview {
                 
                 if(finalResult.absoluteString == url.absoluteString) {
                     
-                    dispatch_async(dispatch_get_main_queue()){
+                    dispatch_async(dispatch_get_main_queue()) {
                         
                         completion(url)
                         
@@ -145,7 +138,7 @@ extension SwiftLinkPreview {
                 
             } else {
                 
-                dispatch_async(dispatch_get_main_queue()){
+                dispatch_async(dispatch_get_main_queue()) {
                     
                     completion(url)
                     
@@ -175,20 +168,26 @@ extension SwiftLinkPreview {
                 
             } else {
                 
+                let sourceUrl = url.absoluteString.hasPrefix("http://") || url.absoluteString.hasPrefix("https://") ? url : NSURL(string: "http://\(url)")
+                
                 do {
                     
-                    let htmlCode = try String(contentsOfURL: url).extendedTrim
+                    // Try to get the page with its default enconding
+                    var source = try String(contentsOfURL: sourceUrl!).extendedTrim
                     
-                    self.crawlMetaTags(htmlCode)
-                    self.crawlTitle(htmlCode)
-                    self.crawlDescription(htmlCode)
-                    self.crawlImages(htmlCode)
-                    
-                    completion()
+                    dispatch_async(dispatch_get_main_queue()) {
+                        
+                        source = self.cleanSource(source)
+                        
+                        self.performPageCrawling(source)
+                        
+                        completion()
+                        
+                    }
                     
                 } catch _ as NSError {
                     
-                    onError(PreviewError(type: .ParseError, url: url.absoluteString))
+                    self.tryAnotherEnconding(sourceUrl!, encodingArray: String.availableStringEncodings(), completion: completion, onError: onError)
                     
                 }
                 
@@ -203,27 +202,118 @@ extension SwiftLinkPreview {
         
     }
     
-    // Extract get canonical URL
-    private func extractCanonicalURL() {
+    // Try to get the page using another available encoding instead the page's own encoding
+    private func tryAnotherEnconding(sourceUrl: NSURL, encodingArray: [NSStringEncoding], completion: () -> (), onError: (PreviewError) -> ()) {
         
-        if var canonicalUrl = Regex.pregMatchFirst((self.result["finalUrl"] as! NSURL).absoluteString, regex: Regex.cannonicalUrlPattern, index: 1) {
+        if encodingArray.isEmpty {
             
-            canonicalUrl = canonicalUrl.replace("http://", with: "").replace("https://", with: "")
-            
-            if let slash = canonicalUrl.rangeOfString("/") {
-                
-                let endIndex = canonicalUrl.startIndex.distanceTo(slash.endIndex)
-                canonicalUrl = canonicalUrl.substring(0, end: endIndex > 1 ? endIndex - 1 : 0)
-                
-            }
-            
-            self.result["canonicalUrl"] = canonicalUrl
+            onError(PreviewError(type: .ParseError, url: url.absoluteString))
             
         } else {
             
-            self.result["canonicalUrl"] = self.result["url"]
+            do {
+                
+                var source = try String(contentsOfURL: sourceUrl, encoding: encodingArray[0]).extendedTrim
+                
+                dispatch_async(dispatch_get_main_queue()) {
+                    
+                    source = self.cleanSource(source)
+                    
+                    self.performPageCrawling(source)
+                    
+                    completion()
+                    
+                }
+                
+            } catch _ as NSError {
+                
+                let availablEncodingArray = encodingArray.filter() { $0 != encodingArray[0] }
+                self.tryAnotherEnconding(sourceUrl, encodingArray: availablEncodingArray, completion: completion, onError: onError)
+                
+            }
             
         }
+        
+    }
+    
+    // Removing unnecessary data from the source
+    private func cleanSource(source: String) -> String {
+        
+        var source = source
+        source = source.deleteTagByPattern(Regex.inlineStylePattern)
+        source = source.deleteTagByPattern(Regex.inlineScriptPattern)
+        source = source.deleteTagByPattern(Regex.linkPattern)
+        source = source.deleteTagByPattern(Regex.scriptPattern)
+        source = source.deleteTagByPattern(Regex.commentPattern)
+        
+        return source
+        
+    }
+    
+    
+    // Perform the page crawiling
+    private func performPageCrawling(htmlCode: String) {
+        
+        var htmlCode = htmlCode
+        
+        self.crawlMetaTags(htmlCode)
+        htmlCode = self.crawlTitle(htmlCode)
+        htmlCode = self.crawlDescription(htmlCode)
+        self.crawlImages(htmlCode)
+        
+    }
+    
+    
+    // Extract canonical URL
+    internal func extractCanonicalURL(finalUrl: NSURL!) -> String {
+        
+        let preUrl: String = finalUrl.absoluteString
+        let url = preUrl
+            .replace("http://", with: "")
+            .replace("https://", with: "")
+            .replace("file://", with: "")
+            .replace("ftp://", with: "")
+        
+        if preUrl != url {
+            
+            if let canonicalUrl = Regex.pregMatchFirst(url, regex: Regex.cannonicalUrlPattern, index: 1) {
+                
+                if(!canonicalUrl.isEmpty) {
+                    
+                    return self.extractBaseUrl(canonicalUrl)
+                    
+                } else {
+                    
+                    return self.extractBaseUrl(url)
+                    
+                }
+                
+            } else {
+                
+                return self.extractBaseUrl(url)
+                
+            }
+            
+        } else {
+            
+            return self.extractBaseUrl(preUrl)
+            
+        }
+        
+    }
+    
+    // Extract base URL
+    private func extractBaseUrl(url: String) -> String {
+        
+        var url = url
+        if let slash = url.rangeOfString("/") {
+            
+            let endIndex = url.startIndex.distanceTo(slash.endIndex)
+            url = url.substring(0, end: endIndex > 1 ? endIndex - 1 : 0)
+            
+        }
+        
+        return url
         
     }
     
@@ -247,13 +337,16 @@ extension SwiftLinkPreview {
                     metatag.rangeOfString("name=\"twitter:\(tag)") != nil ||
                     metatag.rangeOfString("name='twitter:\(tag)") != nil ||
                     metatag.rangeOfString("name=\"\(tag)") != nil ||
-                    metatag.rangeOfString("name='\(tag)") != nil) {
+                    metatag.rangeOfString("name='\(tag)") != nil ||
+                    metatag.rangeOfString("itemprop=\"\(tag)") != nil ||
+                    metatag.rangeOfString("itemprop='\(tag)") != nil) {
                     
                     if((self.result[tag] as! String).isEmpty) {
                         
                         if let value = Regex.pregMatchFirst(metatag, regex: Regex.metatagContentPattern, index: 2) {
                             
-                            self.result[tag] = value.decoded
+                            let value = value.decoded.extendedTrim
+                            self.result[tag] = tag == "image" ? self.addImagePrefixIfNeeded(value) : value
                             
                         }
                         
@@ -268,15 +361,33 @@ extension SwiftLinkPreview {
     }
     
     // Crawl for title if needed
-    internal func crawlTitle(htmlCode: String) {
+    internal func crawlTitle(htmlCode: String) -> String {
         
         if let title: String = self.result["title"] as? String {
             
             if title.isEmpty {
                 
-                if let value = Regex.pregMatchFirst(htmlCode, regex: Regex.tittlePattern, index: 2) {
+                if let value = Regex.pregMatchFirst(htmlCode, regex: Regex.titlePattern, index: 2) {
                     
-                    self.result["title"] = value.decoded
+                    if value.isEmpty {
+                        
+                        if let fromBody: String = self.crawlCode(htmlCode, minimum: SwiftLinkPreview.titleMinimumRelevant) {
+                            
+                            if !fromBody.isEmpty {
+                                
+                                self.result["title"] = fromBody.decoded.extendedTrim
+                                
+                                return htmlCode.replace(fromBody, with: "")
+                                
+                            }
+                            
+                        }
+                        
+                    } else {
+                        
+                        self.result["title"] = value.decoded.extendedTrim
+                        
+                    }
                     
                 }
                 
@@ -284,24 +395,28 @@ extension SwiftLinkPreview {
             
         }
         
+        return htmlCode
+        
     }
     
     // Crawl for description if needed
-    internal func crawlDescription(htmlCode: String) {
+    internal func crawlDescription(htmlCode: String) -> String {
         
         if let description: String = self.result["description"] as? String {
             
             if description.isEmpty {
                 
-                if let value: String = self.crawlCode(htmlCode) {
+                if let value: String = self.crawlCode(htmlCode, minimum: SwiftLinkPreview.decriptionMinimumRelevant) {
                     
-                    self.result["description"] = value
+                    self.result["description"] = value.decoded.extendedTrim
                     
                 }
                 
             }
             
         }
+        
+        return htmlCode
         
     }
     
@@ -322,26 +437,11 @@ extension SwiftLinkPreview {
                         
                         for value in values {
                             
-                            var value = value
-                            
-                            if !value.extendedTrim.isEmpty && !value.hasPrefix("https://") && !value.hasPrefix("http://") && !value.hasPrefix("ftp://") {
-                                
-                                if let host = (self.result["finalUrl"] as! NSURL).host as String! {
-                                    if value.hasPrefix("//") {
-                                        value = (self.result["finalUrl"] as! NSURL).scheme + ":" + value
-                                    }
-                                    else {
-                                        value = (self.result["finalUrl"] as! NSURL).scheme + "://" + host + value
-                                    }
-                                }
-                            }
-                            
-                            imgs.append(value)
+                            imgs.append(self.addImagePrefixIfNeeded(value))
                             
                         }
                         
                         self.result["images"] = imgs
-                        
                         
                         if imgs.count > 0 {
                             
@@ -357,60 +457,99 @@ extension SwiftLinkPreview {
             
         } else {
             
-            self.result["images"] = [mainImage]
+            self.result["images"] = [self.addImagePrefixIfNeeded(mainImage)]
             
         }
+        
+    }
+    
+    // Add prefix image if needed
+    private func addImagePrefixIfNeeded(image: String) -> String {
+        
+        var image = image
+        
+        if let canonicalUrl: String = self.result["canonicalUrl"] as? String {
+            
+            if image.hasPrefix("//") {
+                
+                image = "http:" + image
+                
+            } else if image.hasPrefix("/") {
+                
+                image = "http://" + canonicalUrl + image
+                
+            }
+            
+        }
+        
+        return image
         
     }
     
     // Crawl the entire code
-    internal func crawlCode(content: String) -> String {
+    internal func crawlCode(content: String, minimum: Int) -> String {
         
-        let resultSpan = self.getTagContent("span", content: content)
-        let resultParagraph = self.getTagContent("p", content: content)
-        let resultDiv = self.getTagContent("div", content: content)
-        var result = resultSpan
+        let resultFirstSearch = self.getTagContent("p", content: content, minimum: minimum)
         
-        if (resultParagraph.characters.count >= result.characters.count) {
+        if (!resultFirstSearch.isEmpty) {
             
-            if (resultParagraph.characters.count >= resultDiv.characters.count) {
+            return resultFirstSearch
+            
+        } else {
+            
+            let resultSecondSearch = self.getTagContent("div", content: content, minimum: minimum)
+            
+            if (!resultSecondSearch.isEmpty) {
                 
-                result = resultParagraph
+                return resultSecondSearch
                 
             } else {
                 
-                result = resultDiv
+                let resultThirdSearch = self.getTagContent("span", content: content, minimum: minimum)
+                
+                if (!resultThirdSearch.isEmpty) {
+                    
+                    return resultThirdSearch
+                    
+                } else {
+                    
+                    if (resultThirdSearch.characters.count >= resultFirstSearch.characters.count) {
+                        
+                        if (resultThirdSearch.characters.count >= resultThirdSearch.characters.count) {
+                            
+                            return resultThirdSearch
+                            
+                        } else {
+                            
+                            return resultThirdSearch
+                            
+                        }
+                        
+                    } else {
+                        
+                        return resultFirstSearch
+                        
+                    }
+                    
+                }
+                
                 
             }
             
         }
-        
-        return result
         
     }
     
     // Get tag content
-    private func getTagContent(tag: String, content: String) -> String {
+    private func getTagContent(tag: String, content: String, minimum: Int) -> String {
         
         let pattern = Regex.tagPattern(tag)
-        var result = ""
-        var currentMatch = ""
         
         let index = 2
-        let matches = Regex.pregMatchAll(content, regex: pattern, index: index)
+        let rawMatches = Regex.pregMatchAll(content, regex: pattern, index: index)
         
-        for match in matches {
-            
-            currentMatch = match.extendedTrim.tagsStripped
-            
-            if (currentMatch.characters.count >= SwiftLinkPreview.minimumRelevant) {
-                
-                result = match
-                break
-                
-            }
-            
-        }
+        let matches = rawMatches.filter({ $0.extendedTrim.tagsStripped.characters.count >= minimum })
+        var result = matches.count > 0 ? matches[0] : ""
         
         if result.isEmpty {
             
@@ -422,7 +561,7 @@ extension SwiftLinkPreview {
             
         }
         
-        return result.decoded
+        return result
         
     }