+++ SwiftLinkPreviewTests/BodyTests.swift
@@ -32,7 +32,7 @@ class BodyTests: XCTestCase {
     // MARK: - Span
     func setUpSpan() {
         
-        let metaData =
+        var metaData =
             [
                 Constants.random1: String.randomText(),
                 Constants.random2: String.randomText()
@@ -51,12 +51,12 @@ class BodyTests: XCTestCase {
         template = template.replace(Constants.bodyRandomMiddle, with: String.randomText())
         template = template.replace(Constants.bodyRandomPos, with: String.randomText()).extendedTrim
         
-        slp.resetResult()
-        slp.crawlDescription(template)
+        self.slp.resetResult()
+        self.slp.crawlDescription(template)
         
-        let comparable = (slp.result["description"] as! String)
+        let comparable = (self.slp.result["description"] as! String)
         
-        XCTAssert(comparable == metaData[Constants.random1] || comparable == metaData[Constants.random2])
+        XCTAssert(comparable == metaData[Constants.random1]!.decoded || comparable == metaData[Constants.random2]!.decoded)
         
     }
     
@@ -92,12 +92,12 @@ class BodyTests: XCTestCase {
         template = template.replace(Constants.bodyRandomMiddle, with: String.randomText())
         template = template.replace(Constants.bodyRandomPos, with: String.randomText()).extendedTrim
         
-        slp.resetResult()
-        slp.crawlDescription(template)
+        self.slp.resetResult()
+        self.slp.crawlDescription(template)
         
-        let comparable = (slp.result["description"] as! String)
+        let comparable = (self.slp.result["description"] as! String)
         
-        XCTAssert(comparable == metaData[Constants.random1] || comparable == metaData[Constants.random2])
+        XCTAssert(comparable == metaData[Constants.random1]!.decoded || comparable == metaData[Constants.random2]!.decoded)
         
     }
     
@@ -133,12 +133,12 @@ class BodyTests: XCTestCase {
         template = template.replace(Constants.bodyRandomMiddle, with: String.randomText())
         template = template.replace(Constants.bodyRandomPos, with: String.randomText()).extendedTrim
         
-        slp.resetResult()
-        slp.crawlDescription(template)
+        self.slp.resetResult()
+        self.slp.crawlDescription(template)
         
-        let comparable = (slp.result["description"] as! String)
+        let comparable = (self.slp.result["description"] as! String)
         
-        XCTAssert(comparable == metaData[Constants.random1] || comparable == metaData[Constants.random2])
+        XCTAssert(comparable == metaData[Constants.random1]!.decoded || comparable == metaData[Constants.random2]!.decoded)
         
     }