+++ SwiftLinkPreviewTests/MetaTests.swift
@@ -15,6 +15,7 @@ class MetaTests: XCTestCase {
     // MARK: - Vars
     var twitterTemplate = ""
     var facebookTemplate = ""
+    var itempropTemplate = ""
     var metaTemplate = ""
     let slp = SwiftLinkPreview()
     
@@ -25,6 +26,7 @@ class MetaTests: XCTestCase {
         
         self.twitterTemplate = File.toString(Constants.headMetaTwitter)
         self.facebookTemplate = File.toString(Constants.headMetaFacebook)
+        self.itempropTemplate = File.toString(Constants.headMetaItemprop)
         self.metaTemplate = File.toString(Constants.headMetaMeta)
         
     }
@@ -32,7 +34,7 @@ class MetaTests: XCTestCase {
     // MARK: - Twitter
     func setUpTwitterAndRun() {
         
-        let twitterData =
+        var twitterData =
             [
                 Constants.twitterTitle: String.randomText(),
                 Constants.twitterSite: String.randomUrl(),
@@ -51,12 +53,12 @@ class MetaTests: XCTestCase {
         
         twitterTemplate = twitterTemplate.replace(Constants.bodyRandom, with: String.randomTag()).extendedTrim
         
-        slp.resetResult()
-        slp.crawlMetaTags(twitterTemplate)
+        self.slp.resetResult()
+        self.slp.crawlMetaTags(twitterTemplate)
         
-        XCTAssertEqual((slp.result["title"] as! String), twitterData[Constants.twitterTitle])
-        XCTAssertEqual((slp.result["description"] as! String), twitterData[Constants.twitterDescription])
-        XCTAssertEqual((slp.result["image"] as! String), twitterData[Constants.twitterImageSrc])
+        XCTAssertEqual((self.slp.result["title"] as! String), twitterData[Constants.twitterTitle]!.decoded)
+        XCTAssertEqual((self.slp.result["description"] as! String), twitterData[Constants.twitterDescription]!.decoded)
+        XCTAssertEqual((self.slp.result["image"] as! String), twitterData[Constants.twitterImageSrc])
         
     }
     
@@ -73,7 +75,7 @@ class MetaTests: XCTestCase {
     // MARK: - Facebook
     func setUpFacebookAndRun() {
         
-        let facebookData =
+        var facebookData =
             [
                 Constants.facebookTitle: String.randomText(),
                 Constants.facebookSite: String.randomUrl(),
@@ -92,12 +94,12 @@ class MetaTests: XCTestCase {
         
         facebookTemplate = facebookTemplate.replace(Constants.bodyRandom, with: String.randomTag()).extendedTrim
         
-        slp.resetResult()
-        slp.crawlMetaTags(facebookTemplate)
+        self.slp.resetResult()
+        self.slp.crawlMetaTags(facebookTemplate)
         
-        XCTAssertEqual((slp.result["title"] as! String), facebookData[Constants.facebookTitle])
-        XCTAssertEqual((slp.result["description"] as! String), facebookData[Constants.facebookDescription])
-        XCTAssertEqual((slp.result["image"] as! String), facebookData[Constants.facebookImage])
+        XCTAssertEqual((self.slp.result["title"] as! String), facebookData[Constants.facebookTitle]!.decoded)
+        XCTAssertEqual((self.slp.result["description"] as! String), facebookData[Constants.facebookDescription]!.decoded)
+        XCTAssertEqual((self.slp.result["image"] as! String), facebookData[Constants.facebookImage])
         
     }
     
@@ -111,10 +113,51 @@ class MetaTests: XCTestCase {
         
     }
     
+    // MARK: - Facebook
+    func setUpItempropAndRun() {
+        
+        var itempropData =
+            [
+                Constants.title: String.randomText(),
+                Constants.site: String.randomUrl(),
+                Constants.description: String.randomText(),
+                Constants.image: String.randomImage()
+        ]
+        
+        var itempropTemplate = self.itempropTemplate
+        itempropTemplate = itempropTemplate.replace(Constants.headRandomPre, with: String.randomTag())
+        itempropTemplate = itempropTemplate.replace(Constants.headRandomPos, with: String.randomTag())
+        
+        itempropTemplate = itempropTemplate.replace(Constants.title, with: itempropData[Constants.title]!)
+        itempropTemplate = itempropTemplate.replace(Constants.site, with: itempropData[Constants.site]!)
+        itempropTemplate = itempropTemplate.replace(Constants.description, with: itempropData[Constants.description]!)
+        itempropTemplate = itempropTemplate.replace(Constants.image, with: itempropData[Constants.image]!)
+        
+        itempropTemplate = itempropTemplate.replace(Constants.bodyRandom, with: String.randomTag()).extendedTrim
+        
+        self.slp.resetResult()
+        self.slp.crawlMetaTags(itempropTemplate)
+        
+        XCTAssertEqual((self.slp.result["title"] as! String), itempropData[Constants.title]!.decoded)
+        XCTAssertEqual((self.slp.result["description"] as! String), itempropData[Constants.description]!.decoded)
+        XCTAssertEqual((self.slp.result["image"] as! String), itempropData[Constants.image])
+        
+    }
+    
+    func testItemprop() {
+        
+        for _ in 0 ..< 100 {
+            
+            self.setUpItempropAndRun()
+            
+        }
+        
+    }
+    
     // MARK: - Meta
     func setUpMetaAndRun() {
         
-        let metaData =
+        var metaData =
             [
                 Constants.title: String.randomText(),
                 Constants.site: String.randomUrl(),
@@ -133,12 +176,12 @@ class MetaTests: XCTestCase {
         
         metaTemplate = metaTemplate.replace(Constants.bodyRandom, with: String.randomTag()).extendedTrim
         
-        slp.resetResult()
-        slp.crawlMetaTags(metaTemplate)
+        self.slp.resetResult()
+        self.slp.crawlMetaTags(metaTemplate)
         
-        XCTAssertEqual((slp.result["title"] as! String), metaData[Constants.title])
-        XCTAssertEqual((slp.result["description"] as! String), metaData[Constants.description])
-        XCTAssertEqual((slp.result["image"] as! String), metaData[Constants.image])
+        XCTAssertEqual((self.slp.result["title"] as! String), metaData[Constants.title]!.decoded)
+        XCTAssertEqual((self.slp.result["description"] as! String), metaData[Constants.description]!.decoded)
+        XCTAssertEqual((self.slp.result["image"] as! String), metaData[Constants.image])
         
     }